requirements.txt
gradio==4.26.0
openai==1.19.0
python-dotenv==1.0.1
